refactor: create markdoc utilities and move all functions used internally to a internals file.
